Rename duke to lunaris
