+++ build.gradle
@@ -43,12 +43,12 @@ test {
 }
 
 application {
-    mainClass.set("duke.Launcher")
+    mainClass.set("lunaris.Launcher")
 }
 
 shadowJar {
     //archiveFileName = 'duke.jar'
-    archiveBaseName = "duke"
+    archiveBaseName = "lunaris"
     archiveClassifier = null
     dependsOn("distZip", "distTar")
 }
