+++ src/main/java/lunaris/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package lunaris;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
