+++ src/main/java/lunaris/Launcher.java
@@ -1,4 +1,4 @@
-package duke;
+package lunaris;
 
 import javafx.application.Application;
 
@@ -7,6 +7,6 @@ import javafx.application.Application;
  */
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(Duke.class, args);
+        Application.launch(Lunaris.class, args);
     }
 }
