+++ src/main/java/lunaris/task/Deadline.java
@@ -1,4 +1,4 @@
-package duke.task;
+package lunaris.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
