+++ src/main/java/lunaris/task/ToDo.java
@@ -1,4 +1,4 @@
-package duke.task;
+package lunaris.task;
 
 
 public class ToDo extends Task {
