+++ src/test/java/lunaris/task/DeadlineTest.java
@@ -1,4 +1,4 @@
-package duke.task;
+package lunaris.task;
 
 import org.junit.jupiter.api.Test;
