+++ src/test/java/lunaris/task/ToDoTest.java
@@ -1,4 +1,4 @@
-package duke.task;
+package lunaris.task;
 
 import org.junit.jupiter.api.Test;
